refactor preppers to their own files, and consolidate util methods
